requirements.txt
# MONAI 3D 医学影像分割与分析系统
# 核心依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# MONAI 框架
monai>=1.2.0

# 医学影像处理
nibabel>=5.0.0
SimpleITK>=2.2.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# 可视化
matplotlib>=3.7.0
plotly>=5.15.0

# 训练辅助
pytorch-ignite>=0.4.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
